refactor MoyaNetworkFacade
move Comine extension to class
