--- SCSH_3C_iOS/SCSH_3C_iOS/Network/Moya/MoyaNetworkFacade+Combine.swift
@@ -1,25 +1,0 @@
-//
-//  MoyaNetworkFacade+Combine.swift
-//  SCSH_3C_iOS
-//
-//  Created by 辜敬閎 on 2023/7/24.
-//
-
-import Foundation
-import Moya
-import Combine
-
-extension MoyaNetworkFacade {
-    func fetch<T: MoyaApiInterfaceType>(apiInterface: T) -> AnyPublisher<T.OutputModel, Error> {
-        Future<T.OutputModel, Error> { [weak self] promise in
-            guard let self = self else { return }
-            
-            self.fetch(apiInterface: apiInterface, success: { outputModel in
-                promise(.success(outputModel))
-            }, failure: { error in
-                promise(.failure(error))
-            })
-        }
-        .eraseToAnyPublisher()
-    }
-}
